ml/requirements.txt
torch>=2.4.0
torchvision>=0.19.0
timm>=1.0.0
transformers>=4.45.0
onnx>=1.16.0
onnxruntime-gpu>=1.19.0
psycopg2-binary>=2.9.9
numpy>=1.26.0
Pillow>=10.0.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
tqdm>=4.66.0
